requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

jinja2==3.1.3
python-multipart==0.0.9
itsdangerous==2.1.2

msal==1.28.0

boto3==1.34.34
botocore==1.34.34

cryptography==42.0.5
python-dotenv==1.0.1

apscheduler==3.10.4